requirements.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Validation
pydantic>=2.0.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking (Local)
mlflow>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
loguru>=0.7.0

# Type Hints
typing-extensions>=4.8.0

# Web App / Deployment
streamlit>=1.28.0
